Create geofence objects

+++ app/src/main/java/com/gusta/wakemehome/DetailActivity.java
@@ -150,16 +150,17 @@ public class DetailActivity extends AppCompatActivity {
         String alert = mDetailBinding.clockDetails.alert.getText().toString();
 
         // Show error and abort save if one of the mandatory fields is empty
+        //TODO: Check valid latitude and longitude!
         if (latitudeString.isEmpty() || longitudeString.isEmpty() || radiusString.isEmpty()) {
-            Toast.makeText(getApplicationContext(),R.string.mandatory_fields,Toast.LENGTH_SHORT)
-                    .show();
+            Toast.makeText(getApplicationContext(),
+                    R.string.mandatory_fields,Toast.LENGTH_SHORT).show();
             return;
         }
 
         // Parse numeric fields to their appropriate types
         double latitude = Double.parseDouble(latitudeString);
         double longitude = Double.parseDouble(longitudeString);
-        double radius = Double.parseDouble(radiusString);
+        float radius = Float.parseFloat(radiusString);
 
         // "enabled" field is not shown on this screen - keep current value (if exists)
         boolean enabled = (mViewModel == null) ||
